move provenance type attribute kind to own file
