+++ packages/quicktype-core/src/Transformers.ts
@@ -14,7 +14,7 @@ import { type BaseGraphRewriteBuilder } from "./GraphRewriting";
 import { assert, indentationString, panic } from "./support/Support";
 import { EnumType, PrimitiveType, type Type, type TypeKind, UnionType } from "./Type";
 import { type TypeGraph } from "./Type/TypeGraph";
-import { type TypeRef, derefTypeRef } from "./Type/TypeGraphUtils";
+import { type TypeRef, derefTypeRef } from "./Type/TypeRef";
 
 function debugStringForType(t: Type): string {
     const target = followTargetType(t);
