+++ packages/quicktype-core/src/input/JSONSchemaInput.ts
@@ -49,7 +49,7 @@ import {
     transformedStringTypeTargetTypeKindsMap
 } from "../Type";
 import { type TypeBuilder } from "../Type/TypeBuilder";
-import { type TypeRef } from "../Type/TypeGraph";
+import { type TypeRef } from "../Type/TypeRef";
 
 import { type Input } from "./Inputs";
 import { type JSONSchema, JSONSchemaStore } from "./JSONSchemaStore";
