+++ packages/quicktype-core/src/rewrites/ExpandStrings.ts
@@ -15,7 +15,8 @@ import { type GraphRewriteBuilder } from "../GraphRewriting";
 import { type RunContext } from "../Run";
 import { assert, defined } from "../support/Support";
 import { type PrimitiveType } from "../Type/Type";
-import { type TypeGraph, type TypeRef } from "../Type/TypeGraph";
+import { type TypeGraph } from "../Type/TypeGraph";
+import { type TypeRef } from "../Type/TypeRef";
 import { stringTypesForType } from "../Type/TypeUtils";
 
 const MIN_LENGTH_FOR_ENUM = 10;
